requirements.txt
flax==0.6.7
optax==0.1.3
chex==0.1.5
torch==1.11
tqdm==4.64.0
wandb==0.12.18
dm-tree==0.1.7
jaxtyping==0.0.2
frozendict==2.3.4
transformers==4.26.1
six==1.16.0
pyyaml==6.0
sentencepiece==0.1.96
redis==4.3.4
termcolor==1.1.0
google-cloud-storage==2.5.0
Flask==2.2.1
Flask-Cors==3.0.10
jax==0.4.7
jaxlib==0.4.7
h5py==3.7.0
scipy==1.9.0
sklearn==0.0
openai==0.27.2
scikit-image==0.19.3
dill==0.3.5.1
tyro==0.3.23
rouge-score==0.1.2
gcsfs==2022.8.2
streamlit==1.20.0
sseclient-py==1.7.2
